// src/types/Enums/Descriptions/Fortress/WitchesDescription.ts
export enum WitchesDescription {
  ANDRA = 'Прежде чем позволить Андре, женщине человеческой расы, встать во главе одной из своих армий, Тралосск решил проверить её познания и подверг многочисленным испытаниям. Андра ответила на все вопросы, и королю ничего не оставалось, как принять её на службу.',
  MERIST = 'Мерист считали деревенской знахаркой, но когда обнаружилось, что её магические способности намного превосходят те, которыми должна обладать простая травница, её немедленно призвали на службу трону Таталии.',
  MIRLANDA = 'В дни юности Мирланда во всю баловалась тёмными искусствами, но потом поняла, что этот путь ведёт к самоуничтожению. Она научилась ходить по тонкой грани между добром и злом, постоянно поддерживая баланс между светом и тьмой.',
  ROSIC = 'Когда Розик была девчонкой, она подхватила страшную лихорадку. Старейшины деревни даже не надеялись на её выздоровление. Но она выжила и теперь говорит, что именно лихорадке обязана своими магическими способностями.',
  STYG = 'Стиг была одиннадцатой дочерью Тралосска, короля Таталии. Когда стало очевидно, что при десяти старших наследниках ей никогда не видать трона, Стиг занялась колдовством.',
  TIVA = 'Тива заслужила право называться героем, доказав своим начальникам, что превосходит их учёностью и не имеет равных по уму и сообразительности. Она добилась огромных успехов.',
  VERDISH = 'Вердиш чуть не погибла, когда на её родину напали Властители Подземелий. К счастью, она владела лекарским искусством и смогла залечить смертельные раны, которые получила в битве.',
  VOY = 'Вой зовут ведьмой Ветра, и зачастую морские капитаны приходят к ней с просьбами охранить их в плавании и наслать попутный ветер. Она предпочитает солёный воздух океана смраду родных болот.'
}